+++ CHANGELOG.md
@@ -1,5 +1,26 @@
 # AlpacaScope Changelog
 
+## v2.4.0 - 2022-09-20
+
+Added:
+
+ - Added missing support for NexStar 'h', 'J' and 'm' commands
+ - Added NexStar 'P' command support for GPS, RTC and get device version (1.6)
+ - Improved `--debug` output
+
+Changed:
+
+ - Require Go v1.18+
+ - Many bug fixes in NexStar code for Stellarium Mobile support
+ - NexStar tracking mode now reports correctly
+
+Fixed:
+
+ - Fix decoding of ISO8601 timestamps which include fractional seconds
+    for GetUTCDate
+ - Incorrectly returned Lat twice instead of lat/long 
+ - Error out if unable to connect to telescope
+
 ## v2.3.0 - 2022-07-21
 
 Added: